[MH] [000000] Remove extra file
This importing the rake file in this file causes the rake qa:release:wiki to run twice.  You can see this by running rake -T from a project that uses v2.0.1. You'll see the description duplicated for the rake task.  When you run the task it will run through the rake task twice.

I'm pretty sure this was originally added because we had issues during the deploy process.  But I think the Railtie fixed the issue.